% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSib2BR.R
\name{addSib2BR}
\alias{addSib2BR}
\title{Merge Sibling Summary Data with Birth Records}
\usage{
addSib2BR(BRdata, sibData, by = "femaleID")
}
\arguments{
\item{BRdata}{A dataframe containing birth records.}

\item{sibData}{A dataframe of cleaned and processed sibling data from `mmExtractClean`.}

\item{by}{The matching key column name, defaults to "femaleID".}
}
\value{
A dataframe of birth records enriched with sibling mortality and risk factor information.
}
\description{
This function merges individual-level summary data from `sibatIRlvl` with birth records data. 
It calculates additional variables based on the timing of siblings' deaths relative to birth events 
and aggregates these at the child level. The function aims to enrich birth records with maternal 
and sibling mortality risk factors.
}
\examples{
# Assuming 'BRdata' is your birth records dataset and 'sibData' is cleaned sibling data
enriched_BRdata <- addSib2BR(BRdata, sibData)
}
