% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUnmetNeed.R
\name{addUnmetNeed}
\alias{addUnmetNeed}
\title{Add Unmet Need Variables to Dataset}
\usage{
addUnmetNeed(IRdata)
}
\arguments{
\item{IRdata}{A dataframe containing individual recode (IR) file data from DHS. It must
include specific variables used in the calculation of unmet need and related indicators.
The function checks for these variables and calculates the unmet need based on several
conditions and criteria specific to DHS data.}
}
\value{
The function returns the input dataframe with additional columns related to unmet
need for family planning (`fp_unmet_space`, `fp_unmet_limit`, `fp_unmet_tot`), met need
(`fp_met_space`, `fp_met_limit`, `fp_met_tot`), total demand for contraception
(`fp_demand_space`, `fp_demand_limit`, `fp_demand_tot`), and demand satisfied by any or
modern methods (`fp_demsat_mod`, `fp_demsat_any`).
}
\description{
This function calculates and adds several variables related to unmet need for family planning,
met need, and demand for contraception, based on specific criteria outlined in the DHS
(Demographic and Health Surveys) indicators. It checks for the presence of key variables,
constructs the unmet need variable `v626a` according to a revised definition, and then
calculates additional variables representing different aspects of family planning needs and
demand satisfaction. Note that this function is designed to work with DHS data and expects
certain variables to be present in the dataset.
}
\details{
The function first checks if the `v626a` variable (representing unmet need according
to the revised definition) is present and correctly formatted. If not, it proceeds to calculate
`v626a` based on a series of conditions that consider various aspects of reproductive health and
contraception use. These conditions include whether a woman is using contraception, her
fertility preferences, and her reproductive status. After calculating `v626a`, the function
then calculates additional variables to represent different aspects of family planning needs
and demand satisfaction.

The calculation of unmet need and related variables is complex and based on a series of logical
conditions. It is important to ensure that the input data accurately reflects the variables
and conditions expected by the function. This function does not modify any existing values
in the input dataframe but adds new variables based on the calculations.
}
\examples{
# Assuming `IRdata` is your DHS individual recode dataset with the necessary variables
# IRdata <- read.dta("yourData")
# IRdata_with_unmet_need <- addUnmetNeed(IRdata)

}
