% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferStat.R
\name{bufferStat}
\alias{bufferStat}
\title{Calculate Statistics within a Specified Radius Around Points}
\usage{
bufferStat(points_sf, raster_obj, radius, type = "mean", na.rm = T)
}
\arguments{
\item{points_sf}{An sf object containing points around which buffers will be created.}

\item{raster_obj}{A RasterLayer object from which values will be extracted.}

\item{radius}{Numeric value specifying the radius of the buffer around each point, in the units of the coordinate reference system of `points_sf`.}

\item{type}{Character string specifying the type of statistic to compute within each buffer. Supported types are "mean", "median", "sum", "max", "min", "count", and "mode".}

\item{na.rm}{Logical indicating whether to remove missing values from the raster data before computing statistics. Defaults to TRUE.}
}
\value{
Depending on the `type` parameter, the function returns a vector with the calculated statistic for each buffer. If the `type` is "count" or "mode", the function adjusts to handle raster values accordingly.
}
\description{
This function takes a set of points in an sf object and a raster object, creates buffers
around each point with a specified radius, and then computes various statistical summaries
for the raster values within each buffer.
}
\note{
The function assumes the raster and point spatial data are adequately projected and aligned. Users must ensure that the spatial data used as input follow the same coordinate reference system or are appropriately transformed to align.
}
\examples{
# Load necessary libraries
library(sf)
library(raster)

# Example data (Assuming the data is loaded)
# points_sf <- st_as_sf(data.frame(x = c(1, 2), y = c(2, 3), crs = 4326), coords = c("x", "y"))
# raster_obj <- raster(matrix(1:100, 10, 10))
# radius <- 500  # Example radius

# Calculate mean values within buffers
mean_values <- bufferStat(points_sf, raster_obj, radius, type = "mean")

}
