% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanAN_client.R
\name{cleanAN_client}
\alias{cleanAN_client}
\title{Clean and Recode Antenatal Client Data}
\usage{
cleanAN_client(ANdata)
}
\arguments{
\item{ANdata}{A dataframe containing antenatal care client data with specific survey question
codes expected (e.g., c512 for education level).}
}
\value{
Returns the modified ANdata dataframe with additional variables:
  - edLevel_r: Recoded education level with "Don't know" recoded to NA.
  - readWrite_r: Binary recoding of reading and writing skills.
  - ancVisit_n_obs: Number of antenatal visits as observed (with specific codes recoded to NA).
  - ancVisit_n_resp: Number of antenatal visits as reported by the respondent.
  - firstPregnancy_obs: Observation of whether it's the client's first pregnancy.
  - firstPregnancy_resp: Client's report of whether it's their first pregnancy.
  - weeksPreg_prov: Weeks of pregnancy as provided by the healthcare provider.
  - weeksPreg_card: Weeks of pregnancy according to the health card.
  - moneyDel_enough: Financial readiness for delivery, recoded to binary.
  - ds_knows_count: Count of known danger signs.
  - ds_knows_any: Binary indicator if any danger sign is known.
  - ds_knows_pct: Percentage of known danger signs.
}
\description{
This function preprocesses antenatal client data, including verifying survey phases, recoding several
variables related to educational level, reading and writing skills, number of antenatal visits,
pregnancy information, and financial readiness for delivery. It also evaluates the client's awareness
of various danger signs related to pregnancy and computes indices based on this knowledge.
}
\note{
It is critical to ensure that ANdata contains all the required columns with specific
  coding as used in the function. The function currently checks for compatibility with specific
  survey phases ('AF7', 'HT7', 'MW6', 'NP7', 'TZ7') and issues a warning if other phases are used.
}
\examples{
cleaned_data <- cleanAN_client(ANdata = my_antenatal_data)

}
