% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanBR.R
\name{cleanBR}
\alias{cleanBR}
\title{cleanBR}
\usage{
cleanBR(BRdata, n.cores = NULL)
}
\arguments{
\item{BRdata}{This parameter takes a birth recode file from DHS. It should be in stata format, imported without transforming factors}
}
\description{
This function cleans the birth recode data using code from the DHS repo
}
\examples{

# Load data
br_lb7 <- foreign::read.dta(file.path(dataDir, "LBBR7AFL", "LBBR7AFL.DTA"), convert.factors = F)

# Clean BR data 
br_lb7 <- cleanBR(br_lb7)


}
\author{
J.W. Rozelle
}
