% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanBR_addedChildVars.R
\name{cleanBR_addedChildVars}
\alias{cleanBR_addedChildVars}
\title{Add Calculated Child Variables to Birth Record Data}
\usage{
cleanBR_addedChildVars(BRdata)
}
\arguments{
\item{BRdata}{A dataframe representing birth records, which must contain specific columns like `v224` for the first birth index, and identifiers for mothers (`femaleID`) and children (`childID`).}
}
\value{
The input dataframe with two new columns added:
\itemize{
  \item `first_dob_CMC_b3`: The CMC date of birth for the first child, extracted based on the first birth index (`v224`) and converted to an integer format.
  \item `first_c_dobR`: A date of birth in R format for the first child.
}
}
\description{
This function enhances the birth record dataset by calculating and adding new variables related to children's dates of birth. Specifically, it calculates the CMC (Century Month Code) date of birth for the first child and a refined date of birth for the first child based on specific ID mappings. It relies on auxiliary functions `extractByRC` and `extractByID` for these calculations.
}
\details{
The function requires the `stringr` package for string operations. It uses the `extractByRC` and `extractByID` functions to perform the data extraction and transformation tasks. These auxiliary functions must be defined elsewhere in the package or script. The `extractByRC` function is used to derive the CMC date of birth for the first child based on the index variable `v224` and the `b3` column, applying zero-padding as necessary. The `extractByID` function is tailored to extract and refine the date of birth for the first child by utilizing the identifiers for mothers and children, ensuring accurate linkage and data consistency across records.
}
\examples{
# Assuming `BRdata` is your dataset containing birth records
# with necessary columns like `v224`, `femaleID`, `childID`, etc.
BRdata <- cleanBR(BRdata)
BRdata_enhanced <- cleanBR_addedChildVars(BRdata)

}
