% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFC.R
\name{cleanFC}
\alias{cleanFC}
\title{Clean and Process Facility-Level Data}
\usage{
cleanFC(FCdata)
}
\arguments{
\item{FCdata}{A dataframe containing facility-level data.}
}
\value{
A dataframe with the original data plus new columns for facility IDs, categorized facility types, management authority categories, and indicators of health service availability and readiness. The function adds variables for each facility indicating the presence of various health services, the readiness to provide those services, and categorizations of facilities according to the DHS and country-specific definitions.
}
\description{
This function processes facility-level data by assigning unique facility IDs, categorizing facilities by type and management authority, and calculating various indicators of health service availability and readiness. It supports data from multiple countries with country-specific categorizations for facility types and management authorities. It also computes indicators related to service availability for family planning, antenatal care, child health services, and more.
}
\details{
The function first assigns unique IDs to each facility and categorizes facilities according to their type and management authority, with specific categorizations depending on the country. It then computes indicators for the availability and readiness of services such as family planning, antenatal care, child health services, and others. The readiness indicators are based on the presence of staff, guidelines, equipment, diagnostics, and medicines and commodities necessary to provide these services. The function supports data from Afghanistan (AF7), Haiti (HT7), Malawi (MW6), Nepal (NP7), and Tanzania (TZ7), with specific handling for each country's survey data.
}
\examples{
# Assuming FCdata is your dataset containing facility-level data:
# FCdata_processed <- cleanFC(FCdata)

}
