% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanIR.R
\name{cleanIR}
\alias{cleanIR}
\title{Clean and Process IR Dataset}
\usage{
cleanIR(dataset)
}
\arguments{
\item{dataset}{A data frame containing the IR dataset to be cleaned and processed.}
}
\value{
A data frame with the cleaned and processed IR dataset.
}
\description{
This function processes the given dataset by adding new variables and cleaning existing ones, particularly for demographic and health survey data. It handles missing values, computes new variables, and standardizes certain measures across different datasets.
}
\details{
The `cleanIR` function performs the following tasks:
- Generates unique IDs for each row.
- Formats the interview date.
- Creates a unique ID for each female respondent.
- Extracts and standardizes various interview-related variables (e.g., day, start time, end time, length).
- Creates consistent IDs for fieldworkers and clusters.
- Recodes age and education level variables.
- Computes differences between fieldworker and respondent attributes.
- Handles various indicators related to ever having sex, being in a union, and age differences.
- Recodes wealth index variables.
- Computes child and sibling mortality indicators.
- Processes family planning variables.
- Handles postnatal care indicators for women and newborns.
}
\examples{
\dontrun{
  clean_data <- cleanIR(raw_dataset)
}
}
