% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanIR_selfCareBarriers.R
\name{cleanIR_selfCareBarriers}
\alias{cleanIR_selfCareBarriers}
\title{Clean IR Data for Self-Care Barriers}
\usage{
cleanIR_selfCareBarriers(IRdata)
}
\arguments{
\item{IRdata}{A data frame containing individual respondent (IR) data. 
The data frame must include specific variables (v467a to v467i) that represent survey responses 
to questions about barriers to accessing medical help for oneself.}
}
\value{
A modified version of the input data frame `IRdata` with additional columns: 
\itemize{
  \item `selfMedBarrier_where`: Binary indicator for knowing where to go for medical help. 
  (Currently set to `NA` and not used.)
  \item `selfMedBarrier_permission`: Binary indicator for getting permission to go for medical help.
  \item `selfMedBarrier_money`: Binary indicator for having the money needed for treatment.
  \item `selfMedBarrier_distance`: Binary indicator for distance to health facility being a barrier.
  \item `selfMedBarrier_transport`: Binary indicator for having to take transport. 
  (Currently set to `NA` and not used.)
  \item `selfMedBarrier_alone`: Binary indicator for not wanting to go alone.
  \item `selfMedBarrier_providerFemale`: Binary indicator for concern about no female health provider.
  \item `selfMedBarrier_provider`: Binary indicator for concern about no provider available.
  \item `selfMedBarrier_drugs`: Binary indicator for concern about no drugs available.
  \item `selfMedBarrier_count`: The total count of barriers faced by the respondent.
  \item `selfMedBarrier_any`: Flag indicating whether any self-care barriers are reported.
}
The function currently processes data specifically for respondents from a certain survey wave (e.g., "MW7"), 
and sets most of the new barrier indicators to `NA` initially, later updating them based on the survey responses.
}
\description{
This function preprocesses IRdata, specifically handling variables related to self-care barriers. 
It converts selected survey responses into binary indicators for various barriers to getting 
medical help for oneself, such as knowing where to go, getting permission, financial constraints, 
distance, transportation issues, going alone, gender of the health provider, availability of the 
provider, and drug availability. It then calculates the total count of barriers faced and flags 
if any barriers exist.
}
\details{
The function is designed to work with data from specific survey waves, indicated by the `v000` variable.
It initially sets most of the new barrier indicators to `NA`, updating them based on the respondent's answers 
to relevant questions (v467a to v467i). This preprocessing is crucial for subsequent analysis of barriers to 
self-medication and care within the surveyed population. Note that some indicators are currently not in use 
and are hardcoded to `NA`, reflecting either the unavailability of data or decisions to exclude certain barriers 
from analysis.
}
\examples{
# Assuming `IRdata` is your dataset loaded with appropriate variables (v467a to v467i)
cleanIRdata <- cleanIR_selfCareBarriers(IRdata)
head(cleanIRdata)

}
