% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanKR.R
\name{cleanKR}
\alias{cleanKR}
\title{Clean and Enhance Kids' Recode (KR) Dataset}
\usage{
cleanKR(KRdata)
}
\arguments{
\item{KRdata}{A dataframe representing the Kids' Recode (KR) dataset from DHS, containing variables related to children's health, particularly ARI, fever, and diarrhea indicators.}
}
\value{
The function returns the input dataframe augmented with new variables related to health indicators and care-seeking behaviors. These include indicators for ARI symptoms, fever symptoms, diarrhea symptoms, treatments sought for these conditions from various sources, and variables indicating whether any or all symptoms were treated.
}
\description{
This function cleans and adds several new variables to the Kids' Recode (KR) dataset, focusing on health indicators such as symptoms of ARI, fever, diarrhea, and the care-seeking behavior associated with these conditions. It generates female IDs, processes ARI indicators, fever indicators, diarrhea indicators, treatment sources, and calculates care-seeking behaviors. It assumes specific variable naming conventions aligned with the Demographic and Health Surveys (DHS).
}
\details{
This function is specifically designed for use with DHS Kids' Recode datasets. It expects certain variables to be present and formats them according to DHS standards. The function adds numerous health-related indicators and performs checks to ensure variable integrity, especially for key identifiers like childID and femaleID. The calculation of new indicators is based on DHS guidelines for ARI, fever, and diarrhea, including treatment seeking from different healthcare sources. Note that some variables and calculations may be country-specific and require the dataset to include relevant information.
}
\examples{
# Assuming KRdata is your KR dataset loaded into R
# KRdata <- read.csv("KRdata.csv")
KRdata_cleaned <- cleanKR(KRdata)

}
