% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanKR_chMicro.R
\name{cleanKR_chMicro}
\alias{cleanKR_chMicro}
\title{Clean and Compute Micronutrient Indicators for Children}
\usage{
cleanKR_chMicro(KRdata)
}
\arguments{
\item{KRdata}{A data frame containing the KR dataset with relevant variables for children.}
}
\value{
A data frame with new variables indicating the micronutrient status of children.
}
\description{
This function processes the given dataset to compute various micronutrient indicators 
for children, based on their age and other demographic variables. The computed indicators 
include whether the child received multiple micronutrient powder, iron supplements, vitamin A supplements, 
deworming medication, therapeutic food, supplemental food, and if the child lives in a household with iodized salt.
}
\details{
The function performs the following operations:
\itemize{
  \item Computes the age of the child based on the available data.
  \item Calculates the weight of the child.
  \item Creates binary variables to indicate whether children received specific micronutrients.
}

The new variables created are:
\itemize{
  \item \code{nt_ch_micro_mp}: Children age 6-23 months given multiple micronutrient powder
  \item \code{nt_ch_micro_iron}: Children age 6-59 months given iron supplements
  \item \code{nt_ch_micro_vas}: Children age 6-59 months given vitamin A supplements
  \item \code{nt_ch_micro_dwm}: Children age 6-59 months given deworming medication
  \item \code{nt_ch_micro_iod}: Children age 6-59 months living in households with iodized salt
  \item \code{nt_ch_food_ther}: Children age 6-35 months given therapeutic food
  \item \code{nt_ch_food_supp}: Children age 6-35 months given supplemental food
}
}
\examples{
\dontrun{
# Load necessary packages
library(dplyr)
library(labelled)

# Example usage
cleaned_data <- cleanKR_chMicro(KRdata)
}

}
