% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanKR_iycf.R
\name{cleanKR_iycf}
\alias{cleanKR_iycf}
\title{Clean and Compute Infant and Child Feeding Indicators}
\usage{
cleanKR_iycf(KRdata)
}
\arguments{
\item{KRdata}{A data frame containing the KR dataset with relevant variables for children.}
}
\value{
A data frame with new variables indicating the feeding status and dietary practices of children.
}
\description{
This function processes the given dataset to compute various feeding indicators for infants and children,
including breastfeeding status, dietary diversity, meal frequency, and consumption of specific foods.
}
\details{
The function performs the following operations:
\itemize{
  \item Calculates the weight of the child.
  \item Creates binary variables to indicate various breastfeeding statuses and complementary feeding practices.
  \item Creates variables to indicate the consumption of specific foods in the day/night before the survey.
  \item Computes minimum dietary diversity, meal frequency, and acceptable diet indicators.
}

The new variables created are:
\itemize{
  \item \code{nt_bf_curr}: Currently breastfeeding - last-born under 2 years
  \item \code{nt_bf_status}: Breastfeeding status for last-born child under 2 years
  \item \code{nt_ebf}: Exclusively breastfed - last-born under 6 months
  \item \code{nt_predo_bf}: Predominantly breastfed - last-born under 6 months
  \item \code{nt_ageapp_bf}: Age-appropriately breastfed - last-born under 2 years
  \item \code{nt_food_bf}: Introduced to solid, semi-solid, or soft foods - last-born 6-8 months
  \item \code{nt_bf_cont_1yr}: Continuing breastfeeding at 1 year (12-15 months) - last-born under 2 years
  \item \code{nt_bf_cont_2yr}: Continuing breastfeeding at 2 years (20-23 months) - last-born under 2 years
  \item \code{nt_formula}: Child given infant formula in day/night before survey - last-born under 2 years
  \item \code{nt_milk}: Child given other milk in day/night before survey - last-born under 2 years
  \item \code{nt_liquids}: Child given other liquids in day/night before survey - last-born under 2 years
  \item \code{nt_bbyfood}: Child given fortified baby food in day/night before survey - last-born under 2 years
  \item \code{nt_grains}: Child given grains in day/night before survey - last-born under 2 years
  \item \code{nt_vita}: Child given vitamin A rich food in day/night before survey - last-born under 2 years
  \item \code{nt_frtveg}: Child given other fruits or vegetables in day/night before survey - last-born under 2 years
  \item \code{nt_root}: Child given roots or tubers in day/night before survey - last-born under 2 years
  \item \code{nt_nuts}: Child given legumes or nuts in day/night before survey - last-born under 2 years
  \item \code{nt_meatfish}: Child given meat, fish, shellfish, or poultry in day/night before survey - last-born under 2 years
  \item \code{nt_eggs}: Child given eggs in day/night before survey - last-born under 2 years
  \item \code{nt_dairy}: Child given cheese, yogurt, or other milk products in day/night before survey - last-born under 2 years
  \item \code{nt_solids}: Child given any solid or semisolid food in day/night before survey - last-born under 2 years
  \item \code{nt_fed_milk}: Child given milk or milk products - last-born 6-23 months
  \item \code{nt_mdd}: Child with minimum dietary diversity - last-born 6-23 months
  \item \code{nt_mmf}: Child with minimum meal frequency - last-born 6-23 months
  \item \code{nt_mad}: Child with minimum acceptable diet - last-born 6-23 months
  \item \code{nt_ch_micro_vaf}: Youngest children age 6-23 mos living with mother given Vit A rich food
  \item \code{nt_ch_micro_irf}: Youngest children age 6-23 mos living with mother given iron rich food
}
}
\examples{
\dontrun{
# Load necessary packages
library(dplyr)
library(haven)
library(naniar)
library(labelled)

# Example usage
cleaned_data <- cleanKR_iycf(KRdata)
}

}
