% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanOB.R
\name{cleanOB}
\alias{cleanOB}
\title{Clean and Process OB Data}
\usage{
cleanOB(OBdata)
}
\arguments{
\item{OBdata}{A dataframe containing the OB dataset.}
}
\value{
A dataframe with new variables for linking ID, and indicators for vitamin B12 and vitamin A insufficiency and deficiency.
}
\description{
This function processes the OB dataset by creating a unique linking ID and generating indicators for vitamin B12 and vitamin A insufficiency and deficiency.
}
\details{
The function first creates a unique linking ID by concatenating the cluster number, household number, and respondent's line number. It then generates indicators for the risk of vitamin B12 insufficiency and deficiency, and vitamin A insufficiency and deficiency based on specified thresholds.

The specific labels used for the variables in the OB dataset are:

\describe{
  \item{mnfclust}{Cluster number}
  \item{mnfnumb}{Household number}
  \item{mnfline}{Respondent's line number}
  \item{mn01}{Bar code}
  \item{mn02}{Lab number}
  \item{mn03}{Sample weight}
  \item{mn03a}{Tested for Ferritin}
  \item{mn04}{Ferritin (SF) (2 decimals)}
  \item{mn04ad}{Ferritin Adjusted (SF_adj) (women/children)}
  \item{mn05}{sTfR- soluble transferrin receptor (1 decimal)}
  \item{mn05ad}{sTfR - soluble transferrin receptor - Adjusted (women/children)}
  \item{mn06}{BIS - Body Iron Stores (2 decimals)}
  \item{mn06ad}{BIS - adjusted (2 decimals)}
  \item{mn07}{RBP (retinol-binding protein - 3 decimals)}
  \item{mn07ad}{RBP - Adjusted (women/children)}
  \item{mn08}{CRP (C-reactive protein - 2 decimals)}
  \item{mn09}{AGP (alpha-1-acid glycoprotein - 3 decimals)}
  \item{mn10}{Tested for retinol}
  \item{mn11}{RETINOL (3 decimals)}
  \item{mn11ad}{RETINOL - adjusted (women/children)}
  \item{mn12}{Tested for Vitamin B12}
  \item{mn13}{Vitamin B12}
  \item{mn14}{Tested for UIC (Urinary iodine concentration)}
  \item{mn15}{UIC (1 decimal)}
}
}
\note{
The function assumes that the OB dataset follows the structure outlined in the labels.
}
\examples{
\dontrun{
# Assuming OBdata is already loaded
cleaned_OBdata <- cleanOB(OBdata)
}
}
