% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPR_chNutrition.R
\name{cleanPR_chNutrition}
\alias{cleanPR_chNutrition}
\title{Compute Anthropometry and Anemia Indicators in Children}
\usage{
cleanPR_chNutrition(PRdata)
}
\arguments{
\item{PRdata}{A data frame containing the PR dataset with variables such as \code{hv103}, \code{hc70}, \code{hc72}, \code{hc71}, \code{hc56}, \code{hc1}, and \code{hv005}.}
}
\value{
A data frame with the original variables in \code{PRdata} and new variables for anthropometry and anemia indicators, including:
\item{nt_ch_sev_stunt}{Severely stunted child under 5 years}
\item{nt_ch_stunt}{Stunted child under 5 years}
\item{nt_ch_mean_haz}{Mean z-score for height-for-age for children under 5 years}
\item{nt_ch_sev_wast}{Severely wasted child under 5 years}
\item{nt_ch_wast}{Wasted child under 5 years}
\item{nt_ch_ovwt_ht}{Overweight for height child under 5 years}
\item{nt_ch_mean_whz}{Mean z-score for weight-for-height for children under 5 years}
\item{nt_ch_sev_underwt}{Severely underweight child under 5 years}
\item{nt_ch_underwt}{Underweight child under 5 years}
\item{nt_ch_ovwt_age}{Overweight for age child under 5 years}
\item{nt_ch_mean_waz}{Mean weight-for-age for children under 5 years}
\item{nt_ch_any_anem}{Any anemia - child 6-59 months}
\item{nt_ch_mild_anem}{Mild anemia - child 6-59 months}
\item{nt_ch_mod_anem}{Moderate anemia - child 6-59 months}
\item{nt_ch_sev_anem}{Severe anemia - child 6-59 months}
}
\description{
This function processes the PR dataset to compute various anthropometry and anemia indicators for children under 5 years old.
}
\details{
The function creates new variables based on the following criteria:
\itemize{
  \item Severely stunted: \code{hv103 == 1 & hc70 < -300}
  \item Stunted: \code{hv103 == 1 & hc70 < -200}
  \item Mean height-for-age z-score: \code{hc70 / 100} for valid values
  \item Severely wasted: \code{hv103 == 1 & hc72 < -300}
  \item Wasted: \code{hv103 == 1 & hc72 < -200}
  \item Overweight for height: \code{hv103 == 1 & hc72 > 200}
  \item Mean weight-for-height z-score: \code{hc72 / 100} for valid values
  \item Severely underweight: \code{hv103 == 1 & hc71 < -300}
  \item Underweight: \code{hv103 == 1 & hc71 < -200}
  \item Overweight for age: \code{hv103 == 1 & hc71 > 200}
  \item Mean weight-for-age z-score: \code{hc71 / 100} for valid values
  \item Any anemia: \code{hv103 == 1 & hc1 > 5 & hc1 < 60 & hc56 < 110}
  \item Mild anemia: \code{hv103 == 1 & hc1 > 5 & hc1 < 60 & hc56 > 99 & hc56 < 110}
  \item Moderate anemia: \code{hv103 == 1 & hc1 > 5 & hc1 < 60 & hc56 > 69 & hc56 < 100}
  \item Severe anemia: \code{hv103 == 1 & hc1 > 5 & hc1 < 60 & hc56 < 70}
}

Missing values for each indicator are handled by setting specific values to \code{NA}.
}
\examples{
# Example usage
PRdata <- read.csv("path_to_PR_dataset.csv")
PRdata <- cleanPR_chNutrition(PRdata)

}
