% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPV_toFCLevel.R
\name{cleanPV_toFCLevel}
\alias{cleanPV_toFCLevel}
\title{Aggregate Provider Data to Facility Level}
\usage{
cleanPV_toFCLevel(PVdata)
}
\arguments{
\item{PVdata}{A dataframe of provider-level data, presumably processed by `cleanPV()`, containing variables indicating whether each provider has received specific in-service training.}
}
\value{
A dataframe where each row represents a health facility and columns indicate whether any provider within the facility has received training in specific child health interventions. The function calculates aggregated indicators for each facility, such as the presence of any training in areas like breastfeeding support, infant feeding, IMCI, malaria diagnosis and treatment, and management of pediatric HIV/AIDS.
}
\description{
This function aggregates provider-level data to the facility level, specifically focusing on the availability of in-service training on child health interventions within health facilities. It checks whether at least one provider within a facility has received training in key child health areas such as cold chain management, diagnosis and treatment of ARI and diarrhea, nutrition, breastfeeding, IMCI, and malaria management among others.
}
\details{
Before using this function, the input data must be preprocessed with `cleanPV()` to ensure the presence of necessary variables for in-service training. The function iterates over unique facility IDs, checks for the availability of specific types of in-service training among providers at each facility, and aggregates this information to the facility level. This can help in assessing the overall readiness of health facilities to provide comprehensive child health services based on the training of their staff.
}
\examples{
# Assuming PVdata is your dataset containing provider-level data:
# PVdata_processed <- cleanPV(PVdata)
# FClvlData <- cleanPV_toFCLevel(PVdata_processed)

}
