% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSC_caregiverVars.R
\name{cleanSC_caregiverVars}
\alias{cleanSC_caregiverVars}
\title{Clean and Transform Caregiver Variables}
\usage{
cleanSC_caregiverVars(SCdata)
}
\arguments{
\item{SCdata}{A data frame containing the survey data, expected to include specific variables like `c512`, `c510`, and `c514`.}
}
\value{
The modified data frame with new or transformed variables:
  - `edLevel_r`: Recoded education level, where code `8` is transformed to `NA`.
  - `related_mother`, `related_father`, `related_other`: Binary variables indicating the caregiver's relation to the child (mother, father, or other).
  - `readWrite_cg`: Binary variable indicating whether the caregiver can read and write.
}
\description{
This function modifies a dataset with specific transformations on caregiver-related variables.
It checks the dataset for a specific survey code and updates variables related to education level,
caregiver relationship to child, and literacy status. The function adds warnings if used on unvalidated
data codes.
}
