% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSC_IMCItechquality.R
\name{cleanSC_IMCItechquality}
\alias{cleanSC_IMCItechquality}
\title{Clean and Process SC Data for IMCI Technical Quality Assessment}
\usage{
cleanSC_IMCItechquality(SCdata)
}
\arguments{
\item{SCdata}{A data frame containing the dataset to be processed. This dataset must include specific variables outlined in the function's internal documentation.}
}
\value{
Returns the modified `SCdata` dataset with additional variables related to the IMCI technical quality assessment, including scores and counts for assessments specific to age groups (<2 months, >2 months).
}
\description{
This function processes a dataset (`SCdata`) to assess the Integrated Management of Childhood Illness (IMCI) technical quality. It validates the dataset against specific criteria, checks for the presence of required variables, calculates the IMCI technical quality scores for children aged less than and more than 2 months, and adds several new variables related to IMCI assessment criteria. It issues warnings for unsupported data types and stops execution if required counseling or physical exam variables are missing. This function is specifically validated for datasets with `c000` values of "MW6" or "HT7".
}
\details{
The function first checks for the presence of specific counseling and physical exam variables. If any are missing, the function stops with an error message prompting the user to first process the dataset with `cleanSC_couns()` or `cleanSC_visitActions()`. The function then processes various aspects of the IMCI assessment, such as main symptoms, general danger signs, HIV status inquiries, and feeding assessments, and calculates scores for technical quality based on these criteria. The function has been validated for datasets specifically labeled as "MW6" or "HT7" in the `c000` variable. If a dataset with a different `c000` value is processed, a warning is issued to indicate that the function's accuracy on such data has not been validated.
}
\examples{
# Assuming SCdata is a dataset that has already been loaded and is available in the environment:
SCdata_processed <- cleanSC_IMCItechquality(SCdata)

}
\references{
The source of the IMCI assessment criteria and the technical quality indicators is: DOI: 10.7189/jogh.14.04053.
}
