% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipExtract.R
\name{clipExtract}
\alias{clipExtract}
\title{Clip and Extract Raster Values at Points}
\usage{
clipExtract(rast, points_sf)
}
\arguments{
\item{rast}{A \code{\link[terra]{SpatRaster}} object representing the raster layer.}

\item{points_sf}{An \code{\link[sf]{sf}} object representing the point layer.}
}
\value{
A numeric vector containing the extracted raster values at the locations
        of the points.
}
\description{
This function transforms the coordinates of a point layer to match the CRS of a raster layer,
creates an extent around the transformed points, clips the raster to this extent,
and then extracts the raster values at the original point locations.
}
\examples{
# Load libraries
library(terra)
library(sf)

# Load example raster and points
rast <- rast(system.file("ex/elev.tif", package = "terra"))
points_sf <- st_as_sf(data.frame(x = c(-90, -90.5), y = c(49, 49.5)), coords = c("x", "y"), crs = 4326)

# Perform clipping and extraction
extracted_values <- clipExtract(rast, points_sf)
print(extracted_values)
}
