% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractByID.R
\name{extractByID}
\alias{extractByID}
\title{Extract Variable Value by ID and Index or Lag}
\usage{
extractByID(
  df,
  varName = NULL,
  parentID = "femaleID",
  childID = "childID",
  index = "bord",
  sep = "_",
  lag = NULL,
  indexNumber = NULL
)
}
\arguments{
\item{df}{A dataframe containing the data to be processed.}

\item{varName}{The name of the variable from which to extract values. If `NULL`, the function will return an error.}

\item{parentID}{The name of the column serving as the parent identifier. Default is "femaleID".}

\item{childID}{The name of the column serving as the child identifier. Must be unique for each row.}

\item{index}{The name of the column containing the index or order of children for each parent. Default is "bord".}

\item{sep}{The separator used in constructing reference IDs by concatenating `parentID` and `index`. Default is "_".}

\item{lag}{An optional integer indicating the lag from the current index from which to extract the value. If specified, `indexNumber` should be `NULL`.}

\item{indexNumber}{An optional integer specifying the exact index from which to extract the value. If specified, `lag` should be `NULL`.}
}
\value{
A vector containing the extracted values for the specified variable, corresponding to the given index or lag for each row in the input dataframe. If a match is not found for a constructed reference ID, the function will return `NA` for that entry.
}
\description{
This function extracts the value of a specified variable based on either a direct index number or a lag from the current birth order. It is designed to work with data frames where each row corresponds to a child and rows are grouped by a parent identifier. This function is particularly useful for longitudinal or panel datasets where one needs to reference values from different rows based on a unique ID scheme.
}
\details{
The function first checks for the proper specification of parameters, ensuring either `indexNumber` or `lag` is provided, but not both. It then validates the uniqueness of `childID` values. Subsequent operations include subsetting the data for processing efficiency, verifying ID format, constructing reference IDs based on the given criteria (`indexNumber` or `lag`), and merging the extracted values back into the dataset. The final output is sorted by the original row order to maintain dataset integrity.
}
