% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractByRC.R
\name{extractByRC}
\alias{extractByRC}
\title{Extract Values by Dynamically Constructed Column Names}
\usage{
extractByRC(
  df,
  idx,
  varName,
  varNameIdx_sep = "_",
  lag = 0,
  zeroPad_numDigits = NULL,
  output.class = NULL
)
}
\arguments{
\item{df}{A dataframe from which to extract values.}

\item{idx}{A string specifying the name of the column in `df` that contains the indices used to construct the column names for extraction.}

\item{varName}{The base name of the variables/columns from which to extract values.}

\item{varNameIdx_sep}{Separator to be used between `varName` and the index. Defaults to "_".}

\item{lag}{An integer value to add to the index values found in `idx`. Useful for time-shifted or birth-shifted data extraction. Defaults to 0.}

\item{zeroPad_numDigits}{An optional integer specifying the number of digits for zero padding the index values. If not NULL, indices are zero-padded to this length.}

\item{output.class}{The desired class of the output vector. If specified, can be "numeric" or "integer" to convert the output values to the specified class. If NULL, the class is not altered.}
}
\value{
A vector containing the extracted values from `df`, according to the dynamically constructed column names. If `output.class` is specified, the vector is converted to the specified class.
}
\description{
Dynamically constructs column names from a base name and an index, then extracts values from these columns for each row in a dataframe. It allows for optional zero padding on the indices and lag adjustment.
}
\details{
The function constructs column names by combining `varName`, `varNameIdx_sep`, and the zero-padded (if applicable) indices from the `idx` column, optionally adjusted by `lag`. It then extracts values from these columns for each row in `df`. This is particularly useful for extracting time-series or panel data where the variable of interest changes over time or conditions.
}
\examples{
# Assuming df is your dataframe, idx is the column "timeIndex",
# and you want to extract values from columns named "var_01", "var_02", ..., "var_12"
# with zero padding of 2 digits and a lag of 1:
# extracted_values <- extractByRC(df, "timeIndex", "var", zeroPad_numDigits = 2, lag = 1, output.class = "numeric")

}
