% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineDensity.R
\name{lineDensity}
\alias{lineDensity}
\title{Calculate Road Network Density Around Points}
\usage{
lineDensity(points_sf, roads_sf, radius, uniqueID = "SPAID")
}
\arguments{
\item{points_sf}{An `sf` object containing points.}

\item{roads_sf}{An `sf` object representing the road network.}

\item{radius}{The radius within which to calculate road density, in meters.}
}
\value{
A vector of the same length as points_sf with road density.
}
\description{
This function calculates the density of the road network within a specified
radius around each point in a given set of points. The density is calculated
as the ratio of the total length of the road network within the buffer zone
to the area of the buffer zone. Both the road network and points are provided 
as `sf` objects, and they should be in the same coordinate system.
}
\examples{
# Assuming points_sf and roads_sf are predefined sf objects
result <- lineDensity(points_sf, roads_sf, 1000)

}
