% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb2linestring.R
\name{nb2linestring}
\alias{nb2linestring}
\title{Convert Neighborhood Objects to LINESTRING Spatial Features}
\usage{
nb2linestring(nb_object, sf_object, listw_object = NULL)
}
\arguments{
\item{nb_object}{A neighbor object (e.g., from `poly2nb` function in the `spdep` package).}

\item{sf_object}{An sf object containing the spatial features corresponding to the neighbors.}

\item{listw_object}{Optionally, a listw object containing the spatial weights for each pair
of neighbors. If not provided, weights are set to NA.}
}
\value{
An sf object containing LINESTRING geometries for each neighbor pair,
  with an optional weight attribute if a listw object is provided.
}
\description{
This function takes a neighborhood object, an sf (simple feature) object,
and optionally a listw (spatial weights) object from the `spdep` package
to create LINESTRING geometries representing the connections between neighbors.
Weights from the listw object are incorporated as an attribute if provided.
}
\examples{
\dontrun{
  # Assuming 'nb', 'sf_data', and 'listw' are already defined:
  lines_sf <- nb2linestring(nb, sf_data, listw)
  plot(lines_sf)
}
}
