% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsLengthCalc.R
\name{obsLengthCalc}
\alias{obsLengthCalc}
\title{Calculate Observation Length in Minutes}
\usage{
obsLengthCalc(startVec, endVec, correct = T)
}
\arguments{
\item{startVec}{A character vector representing the start times in "HHMM" format.}

\item{endVec}{A character vector representing the end times in "HHMM" format.}

\item{correct}{A logical value indicating whether to correct instances where the start time
is later than the end time. If TRUE, times are swapped. If FALSE, the function
proceeds without swapping and issues a warning.}
}
\value{
An integer vector of observation lengths in minutes.
}
\description{
This function calculates the difference in minutes between two sets of time points. 
Times are expected to be in "HHMM" format. The function checks for consistency in vector lengths,
and optionally corrects cases where the start time is later than the end time by swapping them.
}
\examples{
startVec <- c("0900", "1430", "1200")
endVec <- c("0930", "1500", "1230")
obsLengthCalc(startVec, endVec)
obsLengthCalc(startVec, c("0915", "1445", "1300"), correct = FALSE)
}
