% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointMetricsInBuffer.R
\name{pointMetricsInBuffer}
\alias{pointMetricsInBuffer}
\title{Calculate Metrics Within Buffers for Point Data}
\usage{
pointMetricsInBuffer(
  hub_sf,
  spoke_sf,
  bufferRadius = 10000,
  bufferGroupVar = NULL,
  metric,
  suffix = "",
  type = "mean"
)
}
\arguments{
\item{hub_sf}{An sf object containing points around which buffers will be created.}

\item{spoke_sf}{An sf object containing points for which metrics will be calculated within the buffers of `hub_sf`.}

\item{bufferRadius}{Numeric, the radius of the buffer(s) in meters. If a single value is provided, 
it applies to all `hub_sf` points. If a named vector is provided, it specifies variable buffer 
radii based on `bufferGroupVar` categories. See the example.}

\item{bufferGroupVar}{Character vector, the name of the variable in `hub_sf` based on which variable buffer 
radii are defined. Required if `bufferRadius` is a named vector.}

\item{metric}{Character vector, names of the variables in `spoke_sf` for which metrics (e.g., mean) 
will be calculated within the buffers.}

\item{suffix}{Character, an optional suffix to append to the metric variable names in the output.}

\item{type}{the type of calculation to be done. Options are "mean", "median", "max", or "min".}
}
\value{
An sf object (`hub_sf`) with new columns added for each of the metrics calculated for the points 
        in `spoke_sf` that fall within the buffer radius of each point in `hub_sf`.
}
\description{
This function calculates specified metrics for `spoke_sf` points within a buffer around `hub_sf` points. 
The buffer radius can be uniform or vary based on a grouping variable. Metrics are calculated for each 
`hub_sf` point and then merged back into the `hub_sf` object.
}
\note{
This function requires the `sf` and `dplyr` packages. Also, ensure that `hub_sf` and `spoke_sf` 
      are in the same coordinate reference system (CRS) and that `bufferRadius` is appropriate for the CRS's 
      units (typically meters for projected CRS).
}
\examples{
# Assuming you have two sf objects, hub_sf and spoke_sf, with appropriate CRS and variables:
# result_sf <- pointMetricsInBuffer(hub_sf, spoke_sf, bufferRadius = c("value1" = 10000, "value2" = 5000), 
                                    bufferGroupVar = "category", metric = c("var1", "var2"))

}
