% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfCareBarriers.R
\name{selfCareBarriers}
\alias{selfCareBarriers}
\title{Barriers for seeking medical care for self}
\format{
A vector with 9 elements.
}
\source{
Variables created as part of cleanIR_selfCareBarriers()
}
\description{
A brief description of what the vector represents. Include any relevant
information about how it was generated or its intended use.
}
\examples{
data(selfMedBarriers_vec)
selfMedBarriers.vec
sum(selfMedBarriers_vec)
}
