% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_kmin.R
\name{which_kmin}
\alias{which_kmin}
\title{Find Indices of the Smallest n Values}
\usage{
which_kmin(x, n)
}
\arguments{
\item{x}{A numeric vector. This vector is the one from which the smallest values
are sought.}

\item{n}{An integer value specifying the number of smallest values to find. This
number should not exceed the length of `x`.}
}
\value{
Returns a vector of indices corresponding to the smallest `n` values
  in the vector `x`.
}
\description{
This function returns the indices of the smallest `n` values in a numeric vector.
It sorts the indices of the vector by the values in ascending order and then selects
the first `n` indices. This is similar to `which.min()` but for multiple values.
}
\examples{
vec <- c(10, 20, 3, 5, 8)
which_kmin(vec, 3)

}
