% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varVecs.R
\name{varVecs}
\alias{varVecs}
\title{A list of Variable vectors}
\format{
A list with the following vectors:
\describe{
  \item{IMCI_lt2m_Vars}{Variables used for IMCI among children 2 months old or less}
  \item{IMCI_gt2m_Vars}{Variables used for IMCI among children more than 2 months old}
  \item{counsSCVars}{Variables used in counseling caretakers for sick children}
  \item{peSCVars}{Variables regarding the physical exam of a sick child.}
  ... and so on for each vector ...
}
}
\source{
Variables created as part of cleanIR_selfCareBarriers()
}
\description{
A list of Variable vectors
}
